Move multiplayer API to top-level jsPsych.multiplayer namespace
Per maintainer suggestion on the PR: the multiplayer methods were
flat-merged into jsPsych.pluginAPI, putting generic names (get, push,
connect, wait, ...) into a shared namespace and requiring an
Object.defineProperty workaround to keep participantId live across the
Object.assign copy. The API is now its own stateful module on the
JsPsych instance, mirroring jsPsych.data:

- Move plugin-api/MultiplayerAPI.ts to modules/multiplayer/ unchanged
- Instantiate jsPsych.multiplayer in the JsPsych constructor; remove
  MultiplayerAPI from the joint pluginAPI object and delete the
  participantId defineProperty workaround
- Re-export GroupSessionData, MultiplayerAdapter, Unsubscribe, and
  MultiplayerTimeoutError from the new module path (public names from
  jspsych are unchanged)
- Rename all call sites in the sync plugin, JATOS adapter doc comment,
  docs, and examples: jsPsych.pluginAPI.* -> jsPsych.multiplayer.*
- Move the API tests to tests/multiplayer/
- Rewrite the changeset to describe jsPsych.multiplayer.* as a state
  primitive (push/getAll/subscribe) plus conveniences and lifecycle
