+++ packages/jspsych/src/index.ts
@@ -66,5 +66,5 @@ export type { JsPsychPlugin, PluginInfo, TrialType } from "./modules/plugins";
 export { ParameterType } from "./modules/plugins";
 export type { JsPsychExtension, JsPsychExtensionInfo } from "./modules/extensions";
 export { DataCollection } from "./modules/data/DataCollection";
-export type { GroupSessionData, MultiplayerAdapter, Unsubscribe } from "./modules/plugin-api";
-export { MultiplayerTimeoutError } from "./modules/plugin-api";
+export type { GroupSessionData, MultiplayerAdapter, Unsubscribe } from "./modules/multiplayer";
+export { MultiplayerTimeoutError } from "./modules/multiplayer";
