+++ packages/jspsych/src/modules/plugin-api/index.ts
@@ -3,12 +3,8 @@ import autoBind from "auto-bind";
 import { JsPsych } from "../../JsPsych";
 import { KeyboardListenerAPI } from "./KeyboardListenerAPI";
 import { MediaAPI } from "./MediaAPI";
-import { MultiplayerAPI } from "./MultiplayerAPI";
 import { SimulationAPI } from "./SimulationAPI";
 import { TimeoutAPI } from "./TimeoutAPI";
-
-export type { GroupSessionData, MultiplayerAdapter, Unsubscribe } from "./MultiplayerAPI";
-export { MultiplayerTimeoutError } from "./MultiplayerAPI";
 
 export function createJointPluginAPIObject(jsPsych: JsPsych) {
   const settings = jsPsych.getInitSettings();
@@ -23,21 +19,12 @@ export function createJointPluginAPIObject(jsPsych: JsPsych) {
     jsPsych.getDisplayContainerElement,
     timeoutAPI.setTimeout.bind(timeoutAPI)
   );
-  const multiplayerAPI = new MultiplayerAPI();
-  const joinedAPI = Object.assign(
+  return Object.assign(
     {},
-    ...[keyboardListenerAPI, timeoutAPI, mediaAPI, simulationAPI, multiplayerAPI].map((object) =>
+    ...[keyboardListenerAPI, timeoutAPI, mediaAPI, simulationAPI].map((object) =>
       autoBind(object)
     )
-  ) as KeyboardListenerAPI & TimeoutAPI & MediaAPI & SimulationAPI & MultiplayerAPI;
-  // Object.assign copies participantId as a null primitive at construction time, so
-  // it never reflects updates made by connect(). Override with a live getter.
-  Object.defineProperty(joinedAPI, "participantId", {
-    get: () => multiplayerAPI.participantId,
-    enumerable: true,
-    configurable: true,
-  });
-  return joinedAPI;
+  ) as KeyboardListenerAPI & TimeoutAPI & MediaAPI & SimulationAPI;
 }
 
 export type PluginAPI = ReturnType<typeof createJointPluginAPIObject>;
