+++ packages/plugin-multiplayer-sync/src/index.ts
@@ -9,7 +9,7 @@ const info = <const>{
     /**
      * Predicate evaluated against the full group session on every update. The trial ends as soon
      * as it returns true. Receives the group session data (keyed by participantId). This is the
-     * same condition you would pass to `jsPsych.pluginAPI.wait()`.
+     * same condition you would pass to `jsPsych.multiplayer.wait()`.
      */
     wait_for: {
       type: ParameterType.FUNCTION,
@@ -84,7 +84,7 @@ type Info = typeof info;
  * push → wait pattern as a single declarative trial so experiments don't have to shoehorn waiting
  * into `call-function` or a `NO_KEYS` keyboard-response trial.
  *
- * Requires a connected multiplayer adapter — call `await jsPsych.pluginAPI.connect(adapter)` before
+ * Requires a connected multiplayer adapter — call `await jsPsych.multiplayer.connect(adapter)` before
  * `jsPsych.run()`. The resolved group session is stored in the trial's `group` data so peer reads
  * and role assignment can happen in a normal `on_finish`.
  *
@@ -97,7 +97,7 @@ class MultiplayerSyncPlugin implements JsPsychPlugin<Info> {
   constructor(private jsPsych: JsPsych) {}
 
   async trial(display_element: HTMLElement, trial: TrialType<Info>) {
-    const api = this.jsPsych.pluginAPI;
+    const api = this.jsPsych.multiplayer;
     display_element.innerHTML = `<div class="jspsych-multiplayer-sync">${trial.message}</div>`;
 
     const start = performance.now();
